Indent sidebar links depending on route level.
The desired indent of any link is computed as reactive declaration.
This desired indent is then assigned to a CSS variable that is used
to style the component's left margin. This only works because style
is scoped to an occurrence of a component in svelte and feels a bit
hacky. This approach is coming from svelte 2 REPL:
https://v2.svelte.dev/repl?version=2.9.9&gist=f041e9a9221be02c1acd82a2712ce3ac

Signed-off-by: Philipp Schneider <philipp@schneiders.space>