Adds `SidebarGroup` & detect active part of tree.
The addition of `SidebarGroup` adds the logical separation between a
single navigation link and the structure of an indented and
collapsible group of navigation links. This allows to keep `Sidebar`
and `SidebarLink` slim.
Additionally the new separation includes the logic in a `SidebarGroup`
whether the group is a parent of the currently active application URL,
which allows to style the active part of the tree differently from the
inactive part of the navigation tree.