Fixes the width calculation for `Sidebar`.
`max()` currently isn't part of the CSS spec. Luckily `max-width`
results in exactly the right behaviour, as it only kicks in once
`min-width` is surpassed. Therefore the width will never be below
320px, even when the `max-width` of 20vw would be less than that.
I overlooked this error, as the `grid-template-columns` rule in
`global.css` resulted in the desired style computation for the `nav`.