Adds select in `App` to simulate route changes.
Previously the route was hardcoded, and it was not possible to see
the `Sidebar` in route change based rerendering.
This new feature immediately uncovers that `setContext()` in `Sidebar`
is not the right tool to pass down changing props to its children.
Other than in react, context is not reactive. A store will be needed.
The `setContext()` implementation in `Sidebar` wouldn't have worked
*even if* context was reactive, since it was only called on component
mount.