Replaces context with writable store.
This commit addresses the bugs uncovered in
56aeeda579e9bb0076cbb305dc49a90cb02d66cd caused by the erroneous
choice to use context by replacing context with a writable store.
The working `activeUrl` change with the select in `App` uncovers an
additional bug: when a route of a collapsed group is set to `active`,
the group is not uncollapsed. The reason is that `SidebarLink`s of a
`SidebarGroup` are not even rendered, so that there's no `SidebarLink`
rendered that could detect via the store change that it's now active
and escalate this with the active event to its `SidebarGroup`. This
is the nail in the coffin of the implementation choice of not
rendering `SidebarLink`s when the group is collapsed. While this fact
was elegant, as it's really in line with reactive JS component
programming, it also causes this bug. I can't think of another
solution than to hide / unhide rather than render / not render.