Enables code-splitting of `Sidebar` out of app.
`App.svelte` now dynamically imports `Sidebar`. Rollup bundles
`main.js` mounting `App.svelte` separately from `Sidebar`, as those
two are set as separate entry points. Rollup now bundles ES modules.

The motivation behind this spit-up is to use this repository both as
container for the actual `Sidebar`, as well as the `App` as demo
playground of the component. But while both is supposed to co-exist
in one repository, the sidebar itself is also supposed to be visibly
modular / independent of the rest of the application, both as
demonstration of bundle size, as well as for easy re-usability by
3rd parties.