Turns `App` into small, resettable playground.
The goal is that the user can try out all props of `Sidebar` live in
an interactive playground mode. This commits is a first step in that
direction.

The user can now not only simulate the `activeUrl` prop, but also
edit the sidebar's displayed navigation `routes` with an input. With
the new `reset` button the user can always reset the rendered
`Sidebar`'s props to the default values.
Note: there's currently no pretty-printing, error handling (e.g. when
entering invalid JSON as `routes`), or form styling. Polishing will
follow once the playground is more complete.
The only nicety in terms of code-style of this commit is that all
props are now passed to `Sidebar` by destructing the `sidebarConfig`,
a mechanism that is also used to reset the user input.