Uses deep copy to (re)set props for `Sidebar`.
As I'm about to add the `theme` prop to `App`'s `sidebarConfig`, it
will be an object with more than one level of nesting. To allow users
to dynamically change the deep object property without mutating the
original value, the code is now using a deep copy of the
`initialSidebarConfig`, which is now also immutable (because it's
frozen).