Fixes passing of prop `onLinkClick` to `sidebar`.
As part of the effort to set up a customisable playground, I
attempted to spread all configuration props from (initial)
`sidebarConfig`. I overlooked that this would break the `onLinkClick`
function, for which it anyway doesn't make a lot of sense to have it
be a customisable prop. Therefore I decided to extract it from the
config object.