Adds light/dark mode switch to site.
If the user uses dark mode, both sidebar and the main page content
use darker background colours.
A downside of the current implementation is that only the main page
content's background colour adapts dynamically. The sidebar's
background colour only updates after a page reload. The reason is that
the implementation has the switch between the light and the dark mode
colours in the initialisation of the sidebar configuration, which is
not reactive. This is an acceptable trade-off in so far, as the case
that a user switches the operating system theme between light and dark
mode while being on the page is not an extremely frequent one.