Moves `getDeepObjectCopy` function position.
The reasoning is that the sidebar configuration setting is now not
interrupted by the definition of a helper function.
I'm not happy about the new placement either, as it now interrupts the
reading flow of sidebar module loading to initialisation. Placing it
below didn't seem like a good option either, as it then would be used
before being defined in the file.