Makes `Sidebar`'s `max-width` prop visible.
Before this commit, the global grid in the body made it impossible for
the `Sidebar` with the given content to grow over the `20vw` value.
I replaced the grid with an easier to handle flex. The secret sauce is
the right mix out of flex-basis, -grow, and -shrink. E.g. without a
small flex basis on the `.route-content` and `#svelte-sidebar`, the
latter never grew.
The change made it necessary to specify a `max-width` in `Sidebar`
when not open, to disallow the element to grow, as it would in the
flexbox by default.