Extracts input[type='number'] into `NumericInput`.
All number inputs refer to `Sidebar` sizing. The inputs on the
playground happens in `input[type='number']`, but need to be parsed to
a number value with unit before they are passed on to `Sidebar` as
prop.
Before this commit, this translation happened `on:input`, which
cluttered the view template. It's now centralised to a reactive
statement.
Naming of values editable on the playground has now also been unified:
All values are attributes of the new `playgroundConfig` object,
mirroring the naming of `initialSidebarConfig`, and `sidebarConfig`.
The latter one is passed as prop for actual rendering by `Sidebar`.