Makes `label` optional for `Select`, `Textarea`.
In 50b39b0ca28ab5cbe587049e677473e50b706456 and
68d7b746bea598bbcab4b483080092c881385761, `Select` and `Textarea`
became independent component, which were supposed to be reusable.
To account for that I intended the label to be optional (which might
not be required for every input usage). However I forgot to declare
the prop as optional by assigning a default value to it.
I failed to see the bug, because all my usages of the components in
fact pass a `label` prop.