Extracts playground to `src/pages/Playground`.
`App` now only contains the main layout, which could be shared across
multiple routes' pages.
The only existing page at the moment is the `Playground`, which now
lives in the new `src/pages/` folder. I moved the `SidebarCodePreview`
to the `playground` folder, as the component is, in contrast to all
components in `src/components/`, not reusable. It is only used on
the playground, and the code specific to that page should be kept in
proximity.
This commit also creates an `src/exampleConfig/` folder, which
contains everything shown on the playground in terms of initial
example routes and sidebar configuration. The previously inlined
`initialSidebarConfig` has been extracted to
`src/exampleConfig/sidebarProps`.