Lets `Playground` respond to window resizings.
* When the user increases or decreases the window width, the sidebar
  is opened / closed in response to the heuristics.
* When the user increases or decreases the minimum sidebar width, the
  sidebar is closed / opened in response to the heuristics.

The implementation is using `<svelte:window/>`. One gotcha on the
svelte API: the first attempt to not bind to the window width, but the
`Playground`s width using ` bind:clientWidth` failed miserably. That
implementation showed the behaviour that the sidebar would always be
re-opened when closed manually. The reason is that the playgrounds
reactive cycle would be triggered even though the window's width
hadn't changed.

This commit also increased the window width threshold at which the
sidebar is open by default, to increase the amount of cases in which
the content is easily readable.