Moves global style to `App`...
... to decrease the number of blocking requests in the render paths.

While HTTP2 sped up loading of more small files, it's still less
optimal for blocking styles to have multiple blocking requests in the
render path then not, as the total stylesheet is still smaller than
the ~14kb that fit into one roundtrip's request payload.

This change has the nice side-effect that now all stiles are minified.
Previously `global.css` was not part of the build pipeline and
therefore not minimised.