Improves 1st loading performance by chunk preload.
The shared chunk `index.js` now no longer is hashed, which allows me
to preload it. This cuts down first loads time nicely.

What I do lose with this commit is the easy to host the shared chunk
with infinite caching enabled, as that would prevent the browser to
ever update the shared chunk.
Including `index.html` in the build, reintroducing the hash, and
automatically adding it to `index.html` would be the optimal solution.