Adds `Readme` component to render `README`...
... and renders `Readme` for routes matching `Readme` sections.
To allow navigation to such routes, this commit changes the example
routes to reference sections from the `README`.

`Readme` loads the repositories `README` that is injected as HTML
string through the md resolving in the bundler, which was added in
c8486ec60f25a38104abbf8286983d8b2d9d8cd4.
`Readme` automatically scrolls to the section referenced in the URL
hash.
`Readme` contains a lot of global styles to style the HTML result of
`README`. If there ever were style collisions, all those global styles
would have to be prefixed with a class only occurring on this page.

This commit also changes:
* Names the example routes so that they are more explanatory of the
  different document use-cases of this project that are mentioned in
  `README`.
* Changes the settings for `serve` so that it responds with
  `index.html` for all routes.
* Removes `activeUrl` from `sidebarProps`, as that is now coming from
  the actual document location.
* Extends `App` so that clicking on a link in the sidebar pushes a
  new browser history entry and that, depending on the window
  location, either `Readme` or the `Playground` are rendered.

The commit introduces a bad behaviour in so far as that clicking on a
sidebar navigation link updates URL and rendered content, but using
the browsers back and forward button triggers no reaction on the page.
My decision not to add a library for routing therefore will cause more
follow-up work in later commits to be a good citizen of the web.