Improves active `NavigationLink` style.
For `NavigationLink`s that are `NavigationGroup`'s links, the caret
is shown next to the group link to (un)collapse the group.
When a `NavigationLink` becomes active, it receives some padding and
the new background color for visual styling. To not reduce the
available with for the link's text and cause new line breaks and
layout jumps, active `NavigationLinks` also receive a negative left
margin to offset the padding.
This negative margin was so big that for active `NavigationGroup`'s
parent `NavigationLinks`, that the caret nearly touched the
`NavigationLink`. This is addressed by this commit.