Adds headings intersection observer to `Readme`.
When the user scrolls through the `Readme`, the URl is now updated
when a heading intersects with the top stripe of the viewport. This
event is interpreted as that the user is reading the section of that
heading.

That the URL is updated in `App` through a binding of the `pathname`
prop automatically also enables `Sidebar` to show the active link
more precisely.